+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - No ticket - Remove mobile number unique constraint
 - TT-1714 - Implement Collaborator Notifications 
 - TT-1841 - Send admin emails for exiting company in signals 
+- TT-1761 - Remove notification hasn't logged-in
 
 ## Breaking changes
 - TT-1538 - AWS-S3 Pass Bucket setup - Pass is required to be setup in target env and set env variable.